Move route types to separate package for organization and to be more consistent with server model
